requirements.txt
Werkzeug==2.0.1
~lask==2.0.1
Flask==2.0.1
numpy==1.20.3
Pillow==8.2.0
protobuf==3.17.3
tensorflow==2.5.0
